Updated layer changing to use ModTap
Layer changing now uses a combination of ModTap and TapDance so that to change from layer 0 to any of the other layers you tap the LOCK key multiple times (2x for layer 1, 3x for layer 2, 4x for layer 3). To return you press and hold the same key on the other layer. This also required an update to the kmk firmware in general.